fix norm errors, leaks, split files
