+++ check_line.c
@@ -6,7 +6,7 @@
 /*   By: sadawi <sadawi@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/25 23:07:44 by sadawi            #+#    #+#             */
-/*   Updated: 2020/03/25 23:08:03 by sadawi           ###   ########.fr       */
+/*   Updated: 2020/04/07 19:10:34 by sadawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
