+++ save_input.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   lem-in.c                                           :+:      :+:    :+:   */
+/*   save_input.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: sadawi <sadawi@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/23 17:18:53 by sadawi            #+#    #+#             */
-/*   Updated: 2020/03/29 13:02:45 by sadawi           ###   ########.fr       */
+/*   Updated: 2020/04/07 19:09:11 by sadawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,9 +39,5 @@ t_farm	*save_input(void)
 	line = save_rooms(&farm);
 	check_room_errors(&farm);
 	save_links(&farm, line);
-	
 	return (farm);
 }
-
-//Remember to free file after everything
-//Maybe should learn more about optimal file reading and parsing?
