+++ save_rooms.c
@@ -6,7 +6,7 @@
 /*   By: sadawi <sadawi@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/29 13:01:18 by sadawi            #+#    #+#             */
-/*   Updated: 2020/04/04 19:07:46 by sadawi           ###   ########.fr       */
+/*   Updated: 2020/04/07 19:06:23 by sadawi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,9 +38,7 @@ int		save_line_room(t_farm **farm, t_room **room, char *line)
 	return (1);
 }
 
-
-//can probably change return value to void?
-int		save_command_room(t_farm **farm, t_room **room, char **line, int cmd)
+void	save_command_room(t_farm **farm, t_room **room, char **line, int cmd)
 {
 	save_line_file(farm, *line);
 	get_next_line(0, line);
@@ -50,16 +48,14 @@ int		save_command_room(t_farm **farm, t_room **room, char **line, int cmd)
 			handle_error("Command not followed by valid room.");
 		*room = (*room)->next;
 	}
-	else
-		if (!(get_line_room(room, *line)))
-			handle_error("Command not followed by valid room.");
+	else if (!(get_line_room(room, *line)))
+		handle_error("Command not followed by valid room.");
 	if (!(*farm)->first)
 		(*farm)->first = *room;
 	if (cmd == START)
 		(*farm)->start = *room;
 	else if (cmd == END)
 		(*farm)->end = *room;
-	return (1);
 }
 
 char	*save_rooms(t_farm **farm)
@@ -73,12 +69,10 @@ char	*save_rooms(t_farm **farm)
 	{
 		if (!check_line_comment(line))
 		{
-			
 			if ((command = check_line_command(line)) > 0)
 				save_command_room(farm, &room, &line, command);
-			else
-				if (!save_line_room(farm, &room, line))
-					return (line);
+			else if (!save_line_room(farm, &room, line))
+				return (line);
 		}
 		save_line_file(farm, line);
 	}
